Changes the calc_charging_times to work with 64-bit numbers
This was the only change required for the part 2 problem

Day 6 complete

+++ src/boat_race.rs
@@ -1,8 +1,8 @@
-pub fn calc_charging_times(time: i32, distance: i32) -> (i32, i32) {
-    let time = time as f32;
-    let distance = distance as f32;
-    let lowest = (time - f32::sqrt(time.powf(2.0) - 4.0 * distance)) / 2.0;
-    let highest = (time + f32::sqrt(time.powf(2.0) - 4.0 * distance)) / 2.0;
+pub fn calc_charging_times(time: i64, distance: i64) -> (i64, i64) {
+    let time = time as f64;
+    let distance = distance as f64;
+    let lowest = (time - f64::sqrt(time.powf(2.0) - 4.0 * distance)) / 2.0;
+    let highest = (time + f64::sqrt(time.powf(2.0) - 4.0 * distance)) / 2.0;
 
-    (lowest.floor() as i32 + 1, highest.ceil() as i32 - 1)
+    (lowest.floor() as i64 + 1, highest.ceil() as i64 - 1)
 }
